requirements.txt
attrs==20.3.0
Brotli==1.0.9
certifi==2020.12.5
chardet==3.0.4
click==6.7
Cython==0.28.2
dash==0.21.0
dash-core-components==0.22.1
dash-html-components==0.10.0
dash-renderer==0.12.1
decorator==4.3.0
Flask==1.0.4
Flask-Compress==1.9.0
gunicorn==20.0.4
idna==2.6
importlib-metadata==3.7.3
ipython-genutils==0.2.0
itsdangerous==1.1.0
Jinja2==2.11.3
jsonschema==3.2.0
jupyter-core==4.7.1
lxml==4.6.2
MarkupSafe==1.1.1
mysql-connector-python==8.0.17
nbformat==4.4.0
numpy==1.14.2
pandas==0.22.0
pandas-datareader==0.6.0
plotly==2.5.1
protobuf==3.15.6
pyrsistent==0.17.3
python-dateutil==2.7.2
pytz==2018.4
pywin32==300
requests==2.18.4
requests-file==1.5.1
requests-ftp==0.3.1
six==1.15.0
traitlets==4.3.3
typing-extensions==3.7.4.3
urllib3==1.22
Werkzeug==0.14.1
wrapt==1.12.1
zipp==3.4.1
